Move outside table definition (#316)
