+++ src/components/network/network-table.js
@@ -11,7 +11,7 @@ import Network from './network';
 import VirtualizedTable from '../util/virtualized-table';
 import Tabs from '@material-ui/core/Tabs';
 import Tab from '@material-ui/core/Tab';
-import { useIntl } from 'react-intl';
+import { FormattedMessage, useIntl } from 'react-intl';
 import InputAdornment from '@material-ui/core/InputAdornment';
 import SearchIcon from '@material-ui/icons/Search';
 import TableCell from '@material-ui/core/TableCell';
@@ -24,13 +24,14 @@ import CheckIcon from '@material-ui/icons/Check';
 import ClearIcon from '@material-ui/icons/Clear';
 
 import ViewColumnIcon from '@material-ui/icons/ViewColumn';
-import { FormattedMessage } from 'react-intl';
 import { SelectOptionsDialog } from '../../utils/dialogs';
 import List from '@material-ui/core/List';
 import ListItem from '@material-ui/core/ListItem';
 import ListItemIcon from '@material-ui/core/ListItemIcon';
 import Checkbox from '@material-ui/core/Checkbox';
 import ListItemText from '@material-ui/core/ListItemText';
+
+import { TABLES_COLUMNS_NAMES, TABLES_DEFINITIONS } from './constants';
 
 const useStyles = makeStyles((theme) => ({
     cell: {
@@ -78,815 +79,12 @@ const NetworkTable = (props) => {
 
     const intl = useIntl();
 
-    const TABLES_DEFINITIONS = {
-        SUBSTATIONS: {
-            index: 0,
-            name: 'Substations',
-            columns: [
-                {
-                    width: 400,
-                    id: 'ID',
-                    dataKey: 'id',
-                },
-                {
-                    width: 200,
-                    id: 'Name',
-                    dataKey: 'name',
-                },
-                {
-                    width: 200,
-                    id: 'Country',
-                    dataKey: 'countryName',
-                },
-            ],
-        },
-
-        VOLTAGE_LEVELS: {
-            index: 1,
-            name: 'VoltageLevels',
-            columns: [
-                {
-                    width: 400,
-                    id: 'ID',
-                    dataKey: 'id',
-                },
-                {
-                    width: 200,
-                    id: 'Name',
-                    dataKey: 'name',
-                },
-                {
-                    width: 400,
-                    id: 'SubstationId',
-                    dataKey: 'substationId',
-                },
-                {
-                    width: 200,
-                    id: 'NominalVoltage',
-                    dataKey: 'nominalVoltage',
-                    numeric: true,
-                    fractionDigits: 0,
-                },
-            ],
-        },
-
-        LINES: {
-            index: 2,
-            name: 'Lines',
-            columns: [
-                {
-                    width: 400,
-                    id: 'ID',
-                    dataKey: 'id',
-                },
-                {
-                    width: 200,
-                    id: 'Name',
-                    dataKey: 'name',
-                },
-                {
-                    width: 400,
-                    id: 'VoltageLevelIdSide1',
-                    dataKey: 'voltageLevelId1',
-                },
-                {
-                    width: 400,
-                    id: 'VoltageLevelIdSide2',
-                    dataKey: 'voltageLevelId2',
-                },
-                {
-                    width: 200,
-                    id: 'ActivePowerSide1',
-                    dataKey: 'p1',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ActivePowerSide2',
-                    dataKey: 'p2',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ReactivePowerSide1',
-                    dataKey: 'q1',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ReactivePowerSide2',
-                    dataKey: 'q2',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-            ],
-        },
-
-        TWO_WINDINGS_TRANSFORMERS: {
-            index: 3,
-            name: 'TwoWindingsTransformers',
-            columns: [
-                {
-                    width: 400,
-                    id: 'ID',
-                    dataKey: 'id',
-                },
-                {
-                    width: 200,
-                    id: 'Name',
-                    dataKey: 'name',
-                },
-                {
-                    width: 400,
-                    id: 'VoltageLevelIdSide1',
-                    dataKey: 'voltageLevelId1',
-                },
-                {
-                    width: 400,
-                    id: 'VoltageLevelIdSide2',
-                    dataKey: 'voltageLevelId2',
-                },
-                {
-                    width: 200,
-                    id: 'ActivePowerSide1',
-                    dataKey: 'p1',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ActivePowerSide2',
-                    dataKey: 'p2',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ReactivePowerSide1',
-                    dataKey: 'q1',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ReactivePowerSide2',
-                    dataKey: 'q2',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 150,
-                    id: 'RatioTap',
-                    dataKey: 'ratioTapChangerPosition',
-                    cellRenderer: (cell) =>
-                        EditableCellRender(
-                            cell,
-                            true,
-                            generateTapRequest('Ratio'),
-                            0
-                        ),
-                },
-                {
-                    width: 150,
-                    id: 'PhaseTap',
-                    dataKey: 'phaseTapChangerPosition',
-                    cellRenderer: (cell) =>
-                        EditableCellRender(
-                            cell,
-                            true,
-                            generateTapRequest('Phase'),
-                            0
-                        ),
-                },
-            ],
-        },
-
-        THREE_WINDINGS_TRANSFORMERS: {
-            index: 4,
-            name: 'ThreeWindingsTransformers',
-            columns: [
-                {
-                    width: 400,
-                    id: 'ID',
-                    dataKey: 'id',
-                },
-                {
-                    width: 200,
-                    id: 'Name',
-                    dataKey: 'name',
-                },
-                {
-                    width: 400,
-                    id: 'VoltageLevelIdSide1',
-                    dataKey: 'voltageLevelId1',
-                },
-                {
-                    width: 400,
-                    id: 'VoltageLevelIdSide2',
-                    dataKey: 'voltageLevelId2',
-                },
-                {
-                    width: 400,
-                    id: 'VoltageLevelIdSide3',
-                    dataKey: 'voltageLevelId3',
-                },
-                {
-                    width: 200,
-                    id: 'ActivePowerSide1',
-                    dataKey: 'p1',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ActivePowerSide2',
-                    dataKey: 'p2',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ActivePowerSide3',
-                    dataKey: 'p3',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ReactivePowerSide1',
-                    dataKey: 'q1',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ReactivePowerSide2',
-                    dataKey: 'q2',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ReactivePowerSide3',
-                    dataKey: 'q3',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 150,
-                    id: 'RatioTap1',
-                    dataKey: 'ratioTapChanger1Position',
-                    cellRenderer: (cell) =>
-                        EditableCellRender(
-                            cell,
-                            true,
-                            generateTapRequest('Ratio', 1),
-                            0
-                        ),
-                },
-                {
-                    width: 150,
-                    id: 'RatioTap2',
-                    dataKey: 'ratioTapChanger2Position',
-                    cellRenderer: (cell) =>
-                        EditableCellRender(
-                            cell,
-                            true,
-                            generateTapRequest('Ratio', 2),
-                            0
-                        ),
-                },
-                {
-                    width: 150,
-                    id: 'RatioTap3',
-                    dataKey: 'ratioTapChanger3Position',
-                    cellRenderer: (cell) =>
-                        EditableCellRender(
-                            cell,
-                            true,
-                            generateTapRequest('Ratio', 3),
-                            0
-                        ),
-                },
-                {
-                    width: 150,
-                    id: 'PhaseTap1',
-                    dataKey: 'phaseTapChanger1Position',
-                    cellRenderer: (cell) =>
-                        EditableCellRender(
-                            cell,
-                            true,
-                            generateTapRequest('Phase', 1),
-                            0
-                        ),
-                },
-                {
-                    width: 150,
-                    id: 'PhaseTap2',
-                    dataKey: 'phaseTapChanger2Position',
-                    cellRenderer: (cell) =>
-                        EditableCellRender(
-                            cell,
-                            true,
-                            generateTapRequest('Phase', 2),
-                            0
-                        ),
-                },
-                {
-                    width: 150,
-                    id: 'PhaseTap3',
-                    numeric: true,
-                    cellRenderer: (cell) =>
-                        EditableCellRender(
-                            cell,
-                            true,
-                            generateTapRequest('Phase', 3),
-                            0
-                        ),
-                },
-            ],
-        },
-
-        GENERATORS: {
-            index: 5,
-            name: 'Generators',
-            columns: [
-                {
-                    width: 400,
-                    id: 'ID',
-                    dataKey: 'id',
-                },
-                {
-                    width: 200,
-                    id: 'Name',
-                    dataKey: 'name',
-                },
-                {
-                    width: 400,
-                    id: 'VoltageLevelId',
-                    dataKey: 'voltageLevelId',
-                },
-                {
-                    width: 200,
-                    id: 'ActivePower',
-                    dataKey: 'p',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ReactivePower',
-                    dataKey: 'q',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'TargetP',
-                    dataKey: 'targetP',
-                    cellRenderer: (cell) =>
-                        EditableCellRender(
-                            cell,
-                            true,
-                            'equipment.setTargetP({})',
-                            1
-                        ),
-                },
-            ],
-        },
-
-        LOADS: {
-            index: 6,
-            name: 'Loads',
-            columns: [
-                {
-                    width: 400,
-                    id: 'ID',
-                    dataKey: 'id',
-                },
-                {
-                    width: 200,
-                    id: 'Name',
-                    dataKey: 'name',
-                },
-                {
-                    width: 200,
-                    id: 'LoadType',
-                    dataKey: 'type',
-                },
-                {
-                    width: 400,
-                    id: 'VoltageLevelId',
-                    dataKey: 'voltageLevelId',
-                },
-                {
-                    width: 200,
-                    id: 'ActivePower',
-                    dataKey: 'p',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ReactivePower',
-                    dataKey: 'q',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ConstantActivePower',
-                    dataKey: 'p0',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ConstantReactivePower',
-                    dataKey: 'q0',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-            ],
-        },
-
-        SHUNT_COMPENSATORS: {
-            index: 7,
-            name: 'ShuntCompensators',
-            columns: [
-                {
-                    width: 400,
-                    id: 'ID',
-                    dataKey: 'id',
-                },
-                {
-                    width: 200,
-                    id: 'Name',
-                    dataKey: 'name',
-                },
-                {
-                    width: 400,
-                    id: 'VoltageLevelId',
-                    dataKey: 'voltageLevelId',
-                },
-                {
-                    width: 200,
-                    id: 'ReactivePower',
-                    dataKey: 'q',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'TargetV',
-                    dataKey: 'targetV',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'TargetDeadband',
-                    dataKey: 'targetDeadband',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-            ],
-        },
-
-        STATIC_VAR_COMPENSATORS: {
-            index: 8,
-            name: 'StaticVarCompensators',
-            columns: [
-                {
-                    width: 400,
-                    id: 'ID',
-                    dataKey: 'id',
-                },
-                {
-                    width: 200,
-                    id: 'Name',
-                    dataKey: 'name',
-                },
-                {
-                    width: 400,
-                    id: 'VoltageLevelId',
-                    dataKey: 'voltageLevelId',
-                },
-                {
-                    width: 200,
-                    id: 'ActivePower',
-                    dataKey: 'p',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ReactivePower',
-                    dataKey: 'q',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'VoltageSetpoint',
-                    dataKey: 'voltageSetpoint',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ReactivePowerSetpoint',
-                    dataKey: 'reactivePowerSetpoint',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-            ],
-        },
-
-        BATTERIES: {
-            index: 9,
-            name: 'Batteries',
-            columns: [
-                {
-                    width: 400,
-                    id: 'ID',
-                    dataKey: 'id',
-                },
-                {
-                    width: 200,
-                    id: 'Name',
-                    dataKey: 'name',
-                },
-                {
-                    width: 400,
-                    id: 'VoltageLevelId',
-                    dataKey: 'voltageLevelId',
-                },
-                {
-                    width: 200,
-                    id: 'ActivePower',
-                    dataKey: 'p',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ReactivePower',
-                    dataKey: 'q',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ConstantActivePower',
-                    dataKey: 'p0',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ConstantReactivePower',
-                    dataKey: 'q0',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-            ],
-        },
-
-        HVDC_LINES: {
-            index: 10,
-            name: 'HvdcLines',
-            columns: [
-                {
-                    width: 400,
-                    id: 'ID',
-                    dataKey: 'id',
-                },
-                {
-                    width: 200,
-                    id: 'Name',
-                    dataKey: 'name',
-                },
-                {
-                    width: 400,
-                    id: 'ConvertersMode',
-                    dataKey: 'convertersMode',
-                },
-                {
-                    width: 400,
-                    id: 'ConverterStationId1',
-                    dataKey: 'converterStationId1',
-                },
-                {
-                    width: 400,
-                    id: 'ConverterStationId2',
-                    dataKey: 'converterStationId2',
-                },
-                {
-                    width: 200,
-                    id: 'R',
-                    dataKey: 'r',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'NominalV',
-                    dataKey: 'nominalV',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 300,
-                    id: 'ActivePowerSetpoint',
-                    dataKey: 'activePowerSetpoint',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'MaxP',
-                    dataKey: 'maxP',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-            ],
-        },
-
-        LCC_CONVERTER_STATIONS: {
-            index: 11,
-            name: 'LccConverterStations',
-            columns: [
-                {
-                    width: 400,
-                    id: 'ID',
-                    dataKey: 'id',
-                },
-                {
-                    width: 200,
-                    id: 'Name',
-                    dataKey: 'name',
-                },
-                {
-                    width: 400,
-                    id: 'VoltageLevelId',
-                    dataKey: 'voltageLevelId',
-                },
-                {
-                    width: 400,
-                    id: 'HvdcLineId',
-                    dataKey: 'hvdcLineId',
-                },
-                {
-                    width: 200,
-                    id: 'ActivePower',
-                    dataKey: 'p',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ReactivePower',
-                    dataKey: 'q',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'PowerFactor',
-                    dataKey: 'powerFactor',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'LossFactor',
-                    dataKey: 'lossFactor',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-            ],
-        },
-
-        VSC_CONVERTER_STATIONS: {
-            index: 12,
-            name: 'VscConverterStations',
-            columns: [
-                {
-                    width: 400,
-                    id: 'ID',
-                    dataKey: 'id',
-                },
-                {
-                    width: 200,
-                    id: 'Name',
-                    dataKey: 'name',
-                },
-                {
-                    width: 400,
-                    id: 'VoltageLevelId',
-                    dataKey: 'voltageLevelId',
-                },
-                {
-                    width: 400,
-                    id: 'HvdcLineId',
-                    dataKey: 'hvdcLineId',
-                },
-                {
-                    width: 200,
-                    id: 'ActivePower',
-                    dataKey: 'p',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ReactivePower',
-                    dataKey: 'q',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'LossFactor',
-                    dataKey: 'lossFactor',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-            ],
-        },
-
-        DANGLING_LINES: {
-            index: 13,
-            name: 'DanglingLines',
-            columns: [
-                {
-                    width: 400,
-                    id: 'ID',
-                    dataKey: 'id',
-                },
-                {
-                    width: 200,
-                    id: 'Name',
-                    dataKey: 'name',
-                },
-                {
-                    width: 400,
-                    id: 'VoltageLevelId',
-                    dataKey: 'voltageLevelId',
-                },
-                {
-                    width: 200,
-                    id: 'UcteXnodeCode',
-                    dataKey: 'ucteXnodeCode',
-                },
-                {
-                    width: 200,
-                    id: 'ActivePower',
-                    dataKey: 'p',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ReactivePower',
-                    dataKey: 'q',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ConstantActivePower',
-                    dataKey: 'p0',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-                {
-                    width: 200,
-                    id: 'ConstantReactivePower',
-                    dataKey: 'q0',
-                    numeric: true,
-                    fractionDigits: 1,
-                },
-            ],
-        },
-    };
-
-    let cols = [];
-    Object.values(TABLES_DEFINITIONS).forEach((table) => {
-        let availableCols = table.columns.map((c) => {
-            return c.id;
-        });
-        cols.push(availableCols);
-    });
-
-    const [listColumnsNames, setListColumnsNames] = useState(cols);
-    const [checked, setChecked] = useState(cols);
+    const [listColumnsNames, setListColumnsNames] = useState(
+        TABLES_COLUMNS_NAMES()
+    );
+    const [checked, setChecked] = useState(TABLES_COLUMNS_NAMES());
     const [listSelectedColumnsNames, setListSelectedColumnsNames] = useState(
-        cols
+        TABLES_COLUMNS_NAMES()
     );
 
     const rowHeight = 48;
@@ -907,24 +105,6 @@ const NetworkTable = (props) => {
         });
     }
 
-    function generateTapRequest(type, leg) {
-        const getLeg = leg !== undefined ? '.getLeg' + leg + '()' : '';
-        return (
-            'tap = equipment' +
-            getLeg +
-            '.get' +
-            type +
-            'TapChanger()\n' +
-            'if (tap.getLowTapPosition() <= {} && {} < tap.getHighTapPosition() ) { \n' +
-            '    tap.setTapPosition({})\n' +
-            // to force update of transformer as sub elements changes like tapChanger are not detected
-            '    equipment.setFictitious(equipment.isFictitious())\n' +
-            '} else {\n' +
-            "throw new Exception('incorrect value')\n" +
-            ' }\n'
-        );
-    }
-
     function commitChanges(rowData) {
         const tab = tabIndex;
         let groovyCr =
@@ -934,7 +114,7 @@ const NetworkTable = (props) => {
             lineEdit[tabIndex].id.replace(/'/g, "\\'") +
             "')\n";
         Object.values(lineEdit[tabIndex].newValues).forEach((cr) => {
-            groovyCr += cr.command.replace(/\{\}/g, cr.value) + '\n';
+            groovyCr += cr.changeCmd.replace(/\{\}/g, cr.value) + '\n';
         });
         requestNetworkChange(props.userId, props.studyName, groovyCr).then(
             (response) => {
@@ -1036,13 +216,13 @@ const NetworkTable = (props) => {
     );
 
     const registerChangeRequest = useCallback(
-        (data, command, value) => {
+        (data, changeCmd, value) => {
             // save original value, dont erase if exists
             if (!lineEdit[tabIndex].oldValues[data.dataKey])
                 lineEdit[tabIndex].oldValues[data.dataKey] =
                     data.rowData[data.dataKey];
             lineEdit[tabIndex].newValues[data.dataKey] = {
-                command: command,
+                changeCmd: changeCmd,
                 value: value,
             };
             data.rowData[data.dataKey] = value;
@@ -1051,7 +231,7 @@ const NetworkTable = (props) => {
     );
 
     const EditableCellRender = useCallback(
-        (cellData, numeric, command, fractionDigit) => {
+        (cellData, numeric, changeCmd, fractionDigit) => {
             return !isLineOnEditMode(cellData.rowIndex) ||
                 cellData.rowData[cellData.dataKey] === undefined ? (
                 defaultCellRender(cellData, numeric, fractionDigit)
@@ -1066,7 +246,7 @@ const NetworkTable = (props) => {
                     onChange={(obj) =>
                         registerChangeRequest(
                             cellData,
-                            command,
+                            changeCmd,
                             obj.target.value
                         )
                     }
@@ -1092,12 +272,22 @@ const NetworkTable = (props) => {
 
     const generateTableColumns = (table) => {
         return table.columns.map((c) => {
-            return {
+            let column = {
                 label: intl.formatMessage({ id: c.id }),
                 headerStyle: { display: showSelectedColumn(c.id) },
                 style: { display: showSelectedColumn(c.id) },
                 ...c,
             };
+            c.changeCmd !== undefined &&
+                (column.cellRenderer = (cell) =>
+                    EditableCellRender(
+                        cell,
+                        c.numeric,
+                        c.changeCmd,
+                        c.fractionDigits
+                    ));
+            delete column.changeCmd;
+            return column;
         });
     };
